Send signal only if mount was done in background

+++ common/util.go
@@ -53,6 +53,7 @@ import (
 )
 
 var RootMount bool
+var ForegroundMount bool
 
 //IsDirectoryMounted is a utility function that returns true if the directory is already mounted using fuse
 func IsDirectoryMounted(path string) bool {
@@ -269,13 +270,15 @@ func ExpandPath(path string) string {
 
 // NotifyMountToParent : Send a signal to parent process about successful mount
 func NotifyMountToParent() error {
-	ppid := syscall.Getppid()
-	if ppid > 1 {
-		if err := syscall.Kill(ppid, syscall.SIGUSR2); err != nil {
-			return err
+	if !ForegroundMount {
+		ppid := syscall.Getppid()
+		if ppid > 1 {
+			if err := syscall.Kill(ppid, syscall.SIGUSR2); err != nil {
+				return err
+			}
+		} else {
+			return fmt.Errorf("failed to get parent pid, received : %v", ppid)
 		}
-	} else {
-		return fmt.Errorf("failed to get parent pid, received : %v", ppid)
 	}
 
 	return nil
